requirements.txt
langchain>=0.2.9
langchain-community>=0.2.9
langchain-text-splitters>=0.2.2
langchain-google-genai>=1.0.7
langchain-huggingface>=0.1.2
sentence-transformers>=3.2.1
chromadb>=0.5.3
pdfplumber>=0.11.0
pypdfium2>=4.30.0
pdf2image>=1.17.0
pytesseract>=0.3.10
streamlit>=1.36.0
fastapi>=0.112.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.1
ragas>=0.1.9
httpx>=0.27.0
tenacity>=8.2.3
numpy>=1.26.0
Pillow>=10.3.0
langchain-groq>=0.1.0
groq>=0.1.0